# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FinTrack API
  description: >
    Production REST API for FinTrack Personal Financial Management and Asset Tracking.
    Implements cash flow tracking, dynamic budgeting, income allocation, customizable
    financial cycles, receivables (loans), investments, and periodic reporting.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker Compose Environment
paths:
...
  /transactions:
    get:
      summary: Query Transactions
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
        - name: type
          in: query
          schema: { type: string, enum: [INCOME, EXPENSE, TRANSFER] }
      responses:
        '200':
          description: Filtered transactions
    post:
      summary: Record Transaction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, type, amount, transaction_date]
              properties:
                account_id: { type: string, format: uuid }
                category_id: { type: string, format: uuid }
                target_account_id: { type: string, format: uuid }
                type: { type: string, enum: [INCOME, EXPENSE, TRANSFER] }
                amount: { type: integer, minimum: 1 }
                transaction_date: { type: string, format: date }
                description: { type: string }
      responses:
        '201':
          description: Transaction recorded
...
  /receivables:
    get:
      summary: List Receivables
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Receivables with balance and status
    post:
      summary: Create Loan/Receivable
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [counterparty, principal]
              properties:
                counterparty: { type: string }
                principal: { type: integer, minimum: 1 }
                source_account_id: { type: string, format: uuid }
                due_date: { type: string, format: date }
                notes: { type: string }
      responses:
        '201':
          description: Receivable created
  /receivables/{id}/payments:
    post:
      summary: Record Loan Repayment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, target_account_id, payment_date]
              properties:
                amount: { type: integer, minimum: 1 }
                target_account_id: { type: string, format: uuid }
                payment_date: { type: string, format: date }
                notes: { type: string }
      responses:
        '201':
          description: Repayment recorded
...
  /reports/monthly:
    get:
      summary: Generate Monthly Financial Statement
      security:
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          schema: { type: integer, minimum: 2000, maximum: 2100 }
        - name: month
          in: query
          schema: { type: integer, minimum: 1, maximum: 12 }
      responses:
        '200':
          description: Consolidated monthly report
        '400':
          description: Invalid year or month
  /reports/cycle:
    get:
      summary: Generate Cycle Financial Statement
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Consolidated cycle report
        '400':
          description: Invalid date format
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT